% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{load_oz}
\alias{load_oz}
\title{Load the australia dataset}
\usage{
load_oz(n = NULL)
}
\arguments{
\item{n}{the number of spectra to return. By default, it returns all 100 spectra in the dataset.}
}
\value{
a SpectraDataFrame
}
\description{
Loads the australia dataset as a SpectraDataFrame
}
\examples{
oz <- load_oz()
summary(oz)

}
\author{
Pierre Roudier
}
