% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_offset}
\alias{plot_offset}
\alias{plot_offset,Spectra-method}
\title{Offset plot of a collection of spectra}
\usage{
plot_offset(x, offset = 1)
}
\arguments{
\item{x}{an object of class \code{"Spectra"} or inheriting from this class}

\item{offset}{Offset between spectra}
}
\description{
Creates an offset plot of a collection of Spectra
}
\examples{
 oz <- load_oz(3)
 plot_offset(oz)
 plot_offset(oz, 0.3)
 plot_offset(oz, 2)
}
\author{
Pierre Roudier
}
