% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot_stack}
\alias{plot_stack}
\alias{plot_stack,Spectra-method}
\title{Stacked plot of a collection of spectra}
\usage{
plot_stack(x)
}
\arguments{
\item{x}{an object of class \code{"Spectra"} or inheriting from this class}
}
\description{
Creates a stacked plot of a collection of Spectra
}
\examples{
 oz <- load_oz(3)
 plot_stack(oz)
}
\author{
Pierre Roudier
}
