% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_summary}
\alias{plot_summary}
\alias{plot_summary,Spectra-method}
\title{Summary plot of a collection of spectra}
\usage{
plot_summary(x, fun = mean, se = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"Spectra"} or inheriting from this class}

\item{fun}{an aggregation function}

\item{se}{if TRUE, plots the standard deviation around the summary spectra
(computed by function as given by \code{fun}). If FALSE, does not plot
dispersion information. If a function, uses this function instead of
\code{sd}.}

\item{...}{additional parameters, currently ignored}
}
\description{
Creates a summary plot of a collection of Spectra
}
\examples{
 oz <- load_oz()
 plot_summary(oz)
}
\author{
Pierre Roudier
}
