% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\docType{methods}
\name{res,numeric-method}
\alias{res,numeric-method}
\alias{res,integer-method}
\alias{res}
\alias{res,Spectra-method}
\alias{res,SpectraDataFrame-method}
\title{Spectral resolution}
\usage{
\S4method{res}{numeric}(x)

\S4method{res}{integer}(x)

\S4method{res}{Spectra}(x)
}
\arguments{
\item{x}{object an object inheriting from \code{Spectra}}
}
\value{
a vector
}
\description{
Returns the spectral resolution of an object
}
\author{
Pierre Roudier \email{pierre.roudier@gmail.com}
}
