% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-methods.R
\name{summary}
\alias{summary}
\alias{summary.Spectra}
\alias{print.summary.Spectra}
\title{Summary}
\usage{
\\method{summary}{Spectra}(object, ...)
}
\arguments{
\item{object}{an object of class \code{Spectra} or \code{SpectraDataFrame}}

\item{...}{Additional arguments passed to \code{summary}}
}
\value{
A \code{"summary.Spectra"} object
}
\description{
Summarize a Spectra* object.
}
\examples{

data(oz)
spectra(oz) <- sr_no ~ ... ~ 350:2500
summary(oz)

}
\author{
Pierre Roudier \url{pierre.roudier@gmail.com}
}
